+++ projects/api/src/contracts/search/schema/response/searchMovieResponseSchema.ts
@@ -1,9 +1,9 @@
 import { movieResponseSchema } from '../../../_internal/response/movieResponseSchema.ts';
-import { int64, z } from '../../../_internal/z.ts';
+import { double, z } from '../../../_internal/z.ts';
 
 /** Zod schema for the search movie response. */
 export const searchMovieResponseSchema = z.object({
-  score: int64(z.number().int()),
+  score: double(z.number()),
   type: z.literal('movie'),
   movie: movieResponseSchema.nullish(),
 });
